Consolidate cell value examples
